# SELU seems to converge faster than ELU