@ikostrikov 2017-06-11
Sorry, I ran multiple experiments with SELU yesterday. I cannot confirm that they converge faster for a3c.